Don't rely on Cloverage to scan for test namespaces
We do our own scanning/loading of tests anyway, the only thing that can happen
is that Cloverage doesn't find anything and so refuses to pass back control to
Kaocha. This is inconvenient since in the case of e.g. Cucumber or CLJS tests
there might be no Clojure namespaces on the test path, but we still want the
test to run.