Get rid of clojure.tools.reader warnings
Cloverage releases a jar which contains an AOT compiled version of a specific
clojure.tools.reader. This seems to have been done in 2015 out of concerns that
people would be using an old and incompatible c.t.r version.

However packing AOT compiled versions of libraries causes other problems.
Notably this has made it impossible to get a c.t.r version that does not give
warnings on Clojure 1.10.

So we released our own jar without AOT. The namespaces have been updated so we
don't clash with other tools.